Move PackageLinks to frontend/model.
