+++ app/lib/shared/urls.dart
@@ -192,40 +192,3 @@ String inferServiceProviderName(String url) {
   }
   return null;
 }
-
-/// The URLs provided by the package's pubspec or inferred from the homepage.
-class PackageLinks {
-  final String homepageUrl;
-  final String documentationUrl;
-  final String repositoryUrl;
-  final String issueTrackerUrl;
-  final String reportIssuesUrl;
-
-  PackageLinks({
-    this.homepageUrl,
-    this.documentationUrl,
-    this.repositoryUrl,
-    this.issueTrackerUrl,
-    this.reportIssuesUrl,
-  });
-
-  factory PackageLinks.infer({
-    String homepageUrl,
-    String documentationUrl,
-    String repositoryUrl,
-    String issueTrackerUrl,
-    String reportIssuesUrl,
-  }) {
-    repositoryUrl ??= inferRepositoryUrl(homepageUrl);
-    issueTrackerUrl ??= inferIssueTrackerUrl(repositoryUrl);
-    reportIssuesUrl ??=
-        inferIssueTrackerUrl(issueTrackerUrl, reportNewIssue: true);
-    return new PackageLinks(
-      homepageUrl: homepageUrl,
-      documentationUrl: documentationUrl,
-      repositoryUrl: repositoryUrl,
-      issueTrackerUrl: issueTrackerUrl,
-      reportIssuesUrl: reportIssuesUrl,
-    );
-  }
-}
